requirement.txt
Cerberus==1.3.7
cfgv==3.4.0
distlib==0.4.0
filelock==3.18.0
identify==2.6.12
nodeenv==1.9.1
platformdirs==4.3.8
pre_commit==4.2.0
PyYAML==6.0.2
ruamel.yaml==0.17.21
setuptools==80.9.0
virtualenv==20.32.0
